+++ packages/paste-codemods/tools/.cache/mappings.json
@@ -261,7 +261,6 @@
   "TextArea": "@twilio-paste/core/textarea",
   "TimePicker": "@twilio-paste/core/time-picker",
   "formatReturnTime": "@twilio-paste/core/time-picker",
-  "AnimatedToast": "@twilio-paste/core/toast",
   "Toast": "@twilio-paste/core/toast",
   "ToastContainer": "@twilio-paste/core/toast",
   "Toaster": "@twilio-paste/core/toast",
@@ -273,7 +272,6 @@
   "Truncate": "@twilio-paste/core/truncate",
   "UserDialog": "@twilio-paste/core/user-dialog",
   "UserDialogContainer": "@twilio-paste/core/user-dialog",
-  "UserDialogContext": "@twilio-paste/core/user-dialog",
   "UserDialogList": "@twilio-paste/core/user-dialog",
   "UserDialogListItem": "@twilio-paste/core/user-dialog",
   "UserDialogSeparator": "@twilio-paste/core/user-dialog",
