+++ packages/paste-core/components/tabs/src/index.tsx
@@ -1,5 +1,16 @@
-export * from "./Tabs";
-export * from "./Tab";
-export * from "./TabList";
-export * from "./TabPanel";
-export * from "./TabPanels";
+import type { TabStateReturn } from "./Tabs";
+
+export { Tabs, useTabState } from "./Tabs";
+export type { TabsProps } from "./Tabs";
+export type { TabStateReturn } from "./Tabs";
+// renaming TabStateReturn to TabStateReturnProps for documenting purposes
+type TabStateReturnProps = TabStateReturn;
+export type { TabStateReturnProps };
+export { Tab } from "./Tab";
+export type { TabProps } from "./Tab";
+export { TabList } from "./TabList";
+export type { TabListProps } from "./TabList";
+export { TabPanel } from "./TabPanel";
+export type { TabPanelProps } from "./TabPanel";
+export { TabPanels } from "./TabPanels";
+export type { TabPanelsProps } from "./TabPanels";
