+++ packages/paste-core/components/textarea/src/index.tsx
@@ -1,1 +1,2 @@
-export * from "./TextArea";
+export { TextArea } from "./TextArea";
+export type { TextAreaProps, TextAreaVariants } from "./TextArea";
