+++ packages/paste-core/components/time-picker/src/index.tsx
@@ -1,2 +1,3 @@
-export * from "./TimePicker";
-export * from "./utils";
+export { TimePicker } from "./TimePicker";
+export type { TimePickerProps } from "./TimePicker";
+export { formatReturnTime } from "./utils";
