+++ packages/paste-core/components/toast/src/index.tsx
@@ -1,5 +1,13 @@
-export * from "./Toast";
-export * from "./ToastContainer";
-export * from "./Toaster";
-export * from "./useToaster";
-export * from "./types";
+export { Toast } from "./Toast";
+export { ToastContainer } from "./ToastContainer";
+export { Toaster } from "./Toaster";
+export { useToaster } from "./useToaster";
+export type {
+  ToastProps,
+  ToasterProps,
+  ToastContainerProps,
+  UseToasterReturnedProps,
+  ToastVariants,
+  ToasterPush,
+  ToasterToast,
+} from "./types";
