+++ packages/paste-core/components/topbar/src/index.tsx
@@ -1,2 +1,4 @@
-export * from "./Topbar";
-export * from "./TopbarActions";
+export { Topbar } from "./Topbar";
+export type { TopbarProps } from "./Topbar";
+export { TopbarActions } from "./TopbarActions";
+export type { TopbarActionsProps } from "./TopbarActions";
