+++ packages/paste-core/components/truncate/src/index.tsx
@@ -3,7 +3,7 @@ import type { BoxProps } from "@twilio-paste/box";
 import type { HTMLPasteProps } from "@twilio-paste/types";
 import * as React from "react";
 
-export interface TruncateProps extends HTMLPasteProps<"span">, Pick<BoxProps, "as"> {
+interface TruncateProps extends HTMLPasteProps<"span">, Pick<BoxProps, "as"> {
   children: NonNullable<React.ReactNode>;
   title: string;
 }
@@ -29,3 +29,4 @@ const Truncate = React.forwardRef<HTMLSpanElement, TruncateProps>(({ children, .
 Truncate.displayName = "Truncate";
 
 export { Truncate };
+export type { TruncateProps };
