+++ packages/paste-core/components/user-dialog/src/UserDialogList.tsx
@@ -1,10 +1,17 @@
 import { Box, safelySpreadBoxProps } from "@twilio-paste/box";
-import type { ListboxPrimitiveInitialState, ListboxPrimitiveStateReturn } from "@twilio-paste/listbox-primitive";
+import type { BoxProps } from "@twilio-paste/box";
+import type {
+  ListboxPrimitiveInitialState,
+  ListboxPrimitiveProps,
+  ListboxPrimitiveStateReturn,
+} from "@twilio-paste/listbox-primitive";
 import { ListboxPrimitive, useListboxPrimitiveState } from "@twilio-paste/listbox-primitive";
+import type { HTMLPasteProps } from "@twilio-paste/types";
 import * as React from "react";
 
-import type { UserDialogListProps, UserDialogListboxProps } from "./types";
-
+interface UserDialogListboxProps extends HTMLPasteProps<"div"> {
+  element?: BoxProps["element"];
+}
 const UserDialogListbox = React.forwardRef<HTMLDivElement, UserDialogListboxProps>(
   ({ element = "USER_DIALOG", ...props }, ref) => {
     return (
@@ -23,6 +30,11 @@ const UserDialogListbox = React.forwardRef<HTMLDivElement, UserDialogListboxProp
 );
 UserDialogListbox.displayName = "UserDialogListbox";
 
+export interface UserDialogListProps extends ListboxPrimitiveProps {
+  children: NonNullable<React.ReactNode>;
+  element?: BoxProps["element"];
+  ["aria-label"]: string;
+}
 export const UserDialogList = React.forwardRef<HTMLDivElement, UserDialogListProps>(({ children, ...props }, ref) => {
   return (
     <ListboxPrimitive {...props} as={UserDialogListbox} ref={ref}>
