+++ packages/paste-core/components/user-dialog/src/index.tsx
@@ -1,8 +1,15 @@
-export * from "./UserDialog";
-export * from "./UserDialogContainer";
-export * from "./UserDialogUserInfo";
-export * from "./UserDialogUserName";
-export * from "./UserDialogUserEmail";
-export * from "./UserDialogList";
-export * from "./UserDialogListItem";
-export * from "./UserDialogSeparator";
+export { UserDialog } from "./UserDialog";
+export type { UserDialogProps } from "./UserDialog";
+export { UserDialogContainer, useUserDialogState } from "./UserDialogContainer";
+export type { UserDialogContainerProps } from "./UserDialogContainer";
+export { UserDialogUserInfo } from "./UserDialogUserInfo";
+export type { UserDialogUserInfoProps } from "./UserDialogUserInfo";
+export { UserDialogUserName } from "./UserDialogUserName";
+export type { UserDialogUserNameProps } from "./UserDialogUserName";
+export { UserDialogUserEmail } from "./UserDialogUserEmail";
+export type { UserDialogUserEmailProps } from "./UserDialogUserEmail";
+export { UserDialogList, useUserDialogListState } from "./UserDialogList";
+export type { UserDialogListProps } from "./UserDialogList";
+export { UserDialogListItem } from "./UserDialogListItem";
+export type { UserDialogListItemProps } from "./UserDialogListItem";
+export { UserDialogSeparator } from "./UserDialogSeparator";
