+++ packages/paste-core/components/visual-picker/src/index.tsx
@@ -1,4 +1,8 @@
-export * from "./VisualPickerRadioGroup";
-export * from "./VisualPickerRadio";
-export * from "./VisualPickerCheckboxGroup";
-export * from "./VisualPickerCheckbox";
+export { VisualPickerRadioGroup } from "./VisualPickerRadioGroup";
+export type { VisualPickerRadioGroupProps } from "./VisualPickerRadioGroup";
+export { VisualPickerRadio } from "./VisualPickerRadio";
+export type { VisualPickerRadioProps } from "./VisualPickerRadio";
+export { VisualPickerCheckboxGroup } from "./VisualPickerCheckboxGroup";
+export type { VisualPickerCheckboxGroupProps } from "./VisualPickerCheckboxGroup";
+export { VisualPickerCheckbox } from "./VisualPickerCheckbox";
+export type { VisualPickerCheckboxProps } from "./VisualPickerCheckbox";
